+++ extensions/twilio/v2/actions/sendSms/config/index.ts
@@ -1,0 +1,1 @@
+export { validateActionFields, fields } from './fields'