+++ extensions/twilio/v2/actions/sendSms/sendSms.ts
@@ -1,56 +1,36 @@
-import { z, ZodError } from 'zod'
+import { ZodError } from 'zod'
 import { fromZodError } from 'zod-validation-error'
-import twilioSdk from '../twilio'
-import { FieldType, type Action, type Field } from '../../../lib/types'
-import { type settings } from '../settings'
-import { Category } from '../../../lib/types/marketplace'
-import { Message, Phone, Settings, validate } from '../validation'
+import twilioSdk from '../../../common/sdk/twilio'
+import { type Action } from '../../../../../lib/types'
+import { type settings } from '../../../settings'
+import { Category } from '../../../../../lib/types/marketplace'
+import { validateSettings } from '../../../settings'
+import { validateActionFields, fields } from './config'
 
-const fields = {
-  message: {
-    id: 'message',
-    label: 'Message',
-    type: FieldType.TEXT,
-    required: true,
-  },
-} satisfies Record<string, Field>
-
-const PatientProfile = z.object({
-  mobile_phone: Phone,
-})
-
-const Patient = z.object({ profile: PatientProfile })
-
-const Schema = z.object({
-  patient: Patient,
-  fields: Message,
-  settings: Settings,
-})
-
-export const patientSmsNotification: Action<typeof fields, typeof settings> = {
-  key: 'patientSmsNotification',
-  title: 'Send SMS to patient',
+export const sendSms: Action<typeof fields, typeof settings> = {
+  key: 'sendSms',
+  title: 'Send SMS',
+  description: 'Send an SMS message.',
   category: Category.COMMUNICATION,
-  description: 'Send an SMS message to the patient enrolled in this care flow.',
   fields,
   onActivityCreated: async (payload, onComplete, onError) => {
     try {
-      const {
-        patient: {
-          profile: { mobile_phone },
-        },
-        fields: { message },
-        settings: { accountSid, authToken, fromNumber },
-      } = validate({ schema: Schema, payload })
+      const { accountSid, authToken, fromNumber } = validateSettings(
+        payload.settings
+      )
+      const { recipient, message } = validateActionFields(payload.fields)
+
       const client = twilioSdk(accountSid, authToken, {
         region: 'IE1',
         accountSid,
       })
+
       await client.messages.create({
         body: message,
         from: fromNumber,
-        to: mobile_phone,
+        to: recipient,
       })
+
       await onComplete()
     } catch (err) {
       if (err instanceof ZodError) {