+++ extensions/twilio/CHANGELOG.md
@@ -1,1 +1,8 @@
 # Twilio changelog
+
+## v2
+
+Compared to v1 (pre-beta release), the extension only has one custom action "Send SMS" instead of two. The Custom Action allows for:
+
+1. Sending a text message to a recipient (phone number) stored in a data point, including patient profile data points.
+2. Sending a text message to a fixed recipient (phone number).